requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.5
    # via langchain
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.5.0
    # via pydantic
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
certifi==2023.7.22
    # via requests
charset-normalizer==3.2.0
    # via
    #   aiohttp
    #   requests
dataclasses-json==0.5.14
    # via langchain
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==2.0.2
    # via sqlalchemy
idna==3.4
    # via
    #   requests
    #   yarl
langchain==0.0.285
    # via -r requirements.in
langsmith==0.0.35
    # via langchain
marshmallow==3.20.1
    # via dataclasses-json
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numexpr==2.8.5
    # via langchain
numpy==1.25.2
    # via
    #   -r requirements.in
    #   langchain
    #   numexpr
packaging==23.1
    # via marshmallow
pydantic==2.3.0
    # via
    #   langchain
    #   langsmith
pydantic-core==2.6.3
    # via pydantic
python-dotenv==1.0.0
    # via -r requirements.in
pyyaml==6.0.1
    # via langchain
requests==2.31.0
    # via
    #   langchain
    #   langsmith
sqlalchemy==2.0.20
    # via langchain
tenacity==8.2.3
    # via langchain
typing-extensions==4.7.1
    # via
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.0.4
    # via requests
yarl==1.9.2
    # via aiohttp
